requirements.txt
Django==4.0
psycopg2-binary==2.9.2
djangorestframework==3.13.1
django-filter==21.1
pillow=12.1.1
djangorestframework-simplejwt
